[FIX] project: open the task form on clicking of task title from kanban
Currently, in the kanban view of the tasks when you click on title of
the task, it will not open the task in the form view.

so in this commit, Removed click event handlers on this widget for
kanban view to ensure that the kanban 'global click' opens the clicked
record, even if the click is done on a title of task.

closes odoo/odoo#70387

Taskid: 2510169
X-original-commit: 2ff5026788edd585cca6d3b53f1f6a65586a4b78
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>