[IMP] project: display commercial_partner_id in project kanban view
Before this commit, when the partner_id in the project, we display the
partner_id selected in this project in the kanban card of it.

This committ displays the commercial_partner_id rather than the
partner_id in the kanban card of each project in the kanban view to be
consist with task's kanban card.

task-2232042

closes #51471