[FIX] website_event: social icons alignment fix
A revamp of the scss of 's_share' in '14.0' has caused an issue with the
class '.o_wevent_social_link' of 'website_event', making a wrong
alignment of the social media icons used in this module, especially in
'event_templates_page_registration'.

For this reason, this commit fixes this issue by setting the display of
the '.o_wevent_social_link' class as an 'inline-flex', vertically and
horizontally centered.

--

task-2508497

closes odoo/odoo#70593

X-original-commit: 49063ee9a72ec03dfca62d28357a0d912b7685bd
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>