[FIX] website_form: prevent input ids from being duplicated
When cloning a website_form field, we rerender the cloned field to
reassign its elements ids and for attributes with a unique ID instead of
the old ones.

Rerendering the list also removes an issue with Selection fields that
were not visible after a clone and a save, as it was usually done after
a _rerenderXML, not executed after a clone.

Also, _renderField docstring was wrong, and its return values
terminology have been replaced with correct one.

task-2485093

X-original-commit: 4c34c9ee96848b3f955732684ab11c18f5b8d244
Part-of: odoo/odoo#75908