[FIX] website: prevent out-of-bounds url autocomplete in linktools
Since the link tools were moved from the modal to the sidebar, the url
autocomplete dropdown stretched out of the bounds of the window.
This fixes that problem while restyling the dropdown to look like the
url pickers we already have in the sidebar. In order to achieve that,
we apply to it the "o_website_ui_autocomplete", used by
UrlPickerUserValueWidget (to which we don't have access in the toolbar).

closes odoo/odoo#70931

X-original-commit: b57275e64b214221339246a2cf00e1b96b03247a
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>