[FW][FIX] crm: always keep master lead probability when merging
Issue
-----

When a lead with an automated probability is merged with other leads and
if its probability is 0, it will be considered as null value and will be
erased by the value of the next lead with a probability > 0

The final lead gets a probability != automated_probability and is now
considered is_automated_probability = False. Therefore automated probability
computation will not be triggered anymore.

Expected behavior
-----------------

Possible use cases
  * if the probability is auto on the master keep the auto probability;
  * if the probability is manual on the master keep that manual value even if
    it is 0 as sales people know their pipe and gave an accurate probability;

In other words: never take the probability from the merged records and always
keep the master one.

Task-2526926
PR odoo#70270

X-Original-Commit: odoo/odoo@cd906edea13aca94736370762b1adfe7b13d7665

closes odoo/odoo#70270

closes odoo/odoo#72073

X-original-commit: 9da59759f933cb12330fdd6a4bd75a855ba86cb9
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>