[FIX] stock_account: return correct empty value for consumable
Creating an account move from a valuation layer is bypassed in
case the product is consumable. Since 8b559b54b5fd, the signature
of _account_entry_move() changed. It no longer return an account move
but a list of values to create everything in batch.
Testing the product type should now return `[]` instead of `False` in
order to be added to the existing list.

closes odoo/odoo#70945

X-original-commit: ecdf7e2f188e16c388ee1da74e4bd03c9387136c
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>